Initial commit of the pkgin test suite.
This is still work in progress, but is able to test a patched version of
pkgin with various operations.  pkgin requires support for PKGIN_DBDIR
and PKG_INSTALL_DIR to be able to set up and operate against local
repositories and package databases.

Expect to see some things move around, and plenty of tests fail at the
moment while fixes to pkgin are slowly merged in.