Add support for per-version exp directories.
Improve compare_output() and file_match() to look for directories that
may override the default, and add expected results files for 0.9 and
0.10 major releases.

Rename skip_before() to skip_if_version() and make it more useful.

Fix a bunch of tests for 0.9.

The test suite now runs to completion on 0.10, and mostly to completion
on 0.9 depending on some nondeterministic output behaviour.