upgrade: Remove old file_match files.
It complicates matters and makes it harder to test arbitrary older
releases that aren't yet supported by the test suite.

Instead, just perform a file_match for the most recent output style, and
for older releases perform enough output_match to be satisfied with the
test.  While historical changes in output can be interesting, it's much
more important we test for features and keep the tests readable.

Also add support for intermediate versions as well as older releases.

The test now runs clean with all pkgin versions 0.6.4 through 22.10.0.